refactor: move error codes
