+++ src/forms/calendar-url-input.tsx
@@ -9,8 +9,8 @@ import { Input } from '@zextras/carbonio-design-system';
 import { t } from '@zextras/carbonio-shell-ui';
 import { isEmpty } from 'lodash';
 
+import { URLErrors } from './error-codes';
 import { InputType } from './types';
-import { URLErrors } from '../soap/types';
 
 export const URLInput = ({
 	value,
