+++ src/soap/types.ts
@@ -4,10 +4,8 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-import { CalendarNameErrorCode, UrlErrorCode } from './error-codes';
+import { CalendarNameErrors, URLErrors } from '../forms/error-codes';
 
-export type URLErrors = Partial<Record<UrlErrorCode, string>>;
-export type CalendarNameErrors = Partial<Record<CalendarNameErrorCode, string>>;
 export type ApiError = {
 	errors: {
 		url?: URLErrors;
